+++ src/main/java/com/app2/engine/entity/app/ReportCreditor.java
@@ -44,4 +44,3 @@ public class ReportCreditor extends BaseEntity {
 
     private String accountNo;
 }
-
