+++ src/main/java/com/app2/engine/entity/app/RequestExtTimeProgress.java
@@ -39,4 +39,3 @@ public class RequestExtTimeProgress extends BaseEntity {
     @JoinColumn(name = "documentProgress")
     private DocumentProgress documentProgress;
 }
-
