--- src/main/java/com/app2/engine/entity/app/RoundRobinAssignedHis.java
@@ -1,31 +1,0 @@
-package com.app2.engine.entity.app;
-
-import com.app2.engine.entity.base.BaseEntity;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-
-@Data
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@EqualsAndHashCode(of = {"id"})
-public class RoundRobinAssignedHis extends BaseEntity {
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "branch")
-    private Branch branch;
-
-    private String branchCenter;
-
-    private String username;
-
-    private String loanType;
-
-}
