+++ src/main/java/com/app2/engine/entity/app/WitnessExamProgress.java
@@ -41,4 +41,3 @@ public class WitnessExamProgress extends BaseEntity {
     @JoinColumn(name = "documentProgress")
     private DocumentProgress documentProgress;
 }
-
