+++ src/main/java/com/app2/engine/repository/ZoneRepository.java
@@ -12,7 +12,7 @@ public interface ZoneRepository extends JpaRepository<Zone, Long> {
 
     Zone findOneByCode(@Param("code")String code);
 
-    @Query("select o  from Zone o where o.code not in :codeList or o.code is null")
+    @Query("select o from Zone o where o.code not in :codeList or o.code is null")
     List<Zone> findByCodeNotIn(@Param("codeList") List<String> codeList);
 }
