+++ src/main/java/com/app2/engine/repository/custom/DocumentProgressRepositoryCustom.java
@@ -7,7 +7,5 @@ import java.util.List;
 import java.util.Map;
 
 public interface DocumentProgressRepositoryCustom {
-    List<Map> findByObj(DocumentProgress documentProgress, List<String> sort, List<String> projection);
     List<DocumentProgress> findByDocumentId(Long documentId);
-    Map getReasonLastProgress(Long documentId);
 }
