--- src/main/java/com/app2/engine/repository/custom/DocumentRepositoryCustom.java
@@ -1,8 +1,0 @@
-package com.app2.engine.repository.custom;
-
-import java.util.List;
-import java.util.Map;
-
-public interface DocumentRepositoryCustom {
-    List<Map> findLsAccountList();
-}
